requirements.txt
streamlit==1.10.0
tensorflow==2.12.0
pillow==9.1.0
numpy==1.21.5
matplotlib==3.5.1
rich>=10.14.0
